@@ -4,8 +4,8 @@ using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using IdentityModel.AspNetCore.AccessTokenManagement;
-using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Options;
 
 public interface IAuthHeaderProvider
 {
@@ -15,12 +15,12 @@ public interface IAuthHeaderProvider
 public class AuthHeaderProvider : IAuthHeaderProvider
 {
     private readonly IClientAccessTokenManagementService tokenService;
-    private readonly IConfiguration config;
+    private readonly AuthConfig config;
     private readonly ILogger<AuthHeaderProvider> log;
 
-    public AuthHeaderProvider(IConfiguration config, IClientAccessTokenManagementService tokenService, ILogger<AuthHeaderProvider> logger)
+    public AuthHeaderProvider(IOptions<AuthConfig> config, IClientAccessTokenManagementService tokenService, ILogger<AuthHeaderProvider> logger)
     {
-        this.config = config;
+        this.config = config.Value;
         log = logger;
         this.tokenService = tokenService;
 
@@ -28,18 +28,18 @@ public class AuthHeaderProvider : IAuthHeaderProvider
         BearerTokenAuthHeader = GetBearerTokenHeader();
     }
 
-    public AuthenticationHeaderValue BasicAuthHeader { get; }
+    private AuthenticationHeaderValue BasicAuthHeader { get; }
 
-    public AuthenticationHeaderValue BearerTokenAuthHeader { get; }
+    private AuthenticationHeaderValue BearerTokenAuthHeader { get; }
 
     public async Task<AuthenticationHeaderValue> GetAuthHeaderAsync(CancellationToken cancelToken = default)
     {
-        if (BasicAuthHeader != null)
+        if (BasicAuthHeader is not null)
         {
             return await Task.FromResult(BasicAuthHeader);
         }
 
-        if (BearerTokenAuthHeader != null)
+        if (BearerTokenAuthHeader is not null)
         {
             return await Task.FromResult(BearerTokenAuthHeader);
         }
@@ -49,29 +49,26 @@ public class AuthHeaderProvider : IAuthHeaderProvider
 
     private async Task<AuthenticationHeaderValue> GetOAuthHeader(CancellationToken cancelToken = default)
     {
-        var oAuthUri = config.GetValue<Uri>("Auth:OAuth:TokenUrl");
-        if (oAuthUri != null)
+        var oAuthUri = config.OAuth.TokenUrl;
+        if (oAuthUri is null)
         {
-            var token = await tokenService.GetClientAccessTokenAsync(cancellationToken: cancelToken);
-
-            if (token == null)
-            {
-                log.LogError("Failed to get oauth token.");
-                return null;
-            }
-
-            log.LogDebug("Using oauth token");
-
-            return new AuthenticationHeaderValue("Bearer", token);
+            return null;
         }
 
-        return null;
+        var token = await tokenService.GetClientAccessTokenAsync(cancellationToken: cancelToken);
+        if (token is null)
+        {
+            log.LogError("Failed to get oauth token.");
+            return null;
+        }
+
+        return new AuthenticationHeaderValue("Bearer", token);
     }
 
     private AuthenticationHeaderValue GetBasicAuthHeader()
     {
-        var username = config.GetValue<string>("Auth:Basic:Username", null);
-        var password = config.GetValue<string>("Auth:Basic:Password", null);
+        var username = config.Basic.Username;
+        var password = config.Basic.Password;
 
         if (!string.IsNullOrWhiteSpace(username) && !string.IsNullOrWhiteSpace(username))
         {
@@ -85,7 +82,7 @@ public class AuthHeaderProvider : IAuthHeaderProvider
 
     private AuthenticationHeaderValue GetBearerTokenHeader()
     {
-        var bearerTokenFromConfig = config.GetValue<string>("Auth:BearerToken");
+        var bearerTokenFromConfig = config.BearerToken;
 
         if (!string.IsNullOrWhiteSpace(bearerTokenFromConfig))
         {
